Synchronize with main repository: https://www.github.com/fDero/Basalt

+++ include/errors/display_utilities.hpp
@@ -1,15 +1,21 @@
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+// AUTHOR: Francesco De Rosa (https://github.com/fDero)                    //
+// LICENSE: MIT (https://github.com/fDero/Basalt/blob/master/LICENSE)      //
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
 #pragma once
+
 #include "errors/tokenization_errors.hpp"
 #include "errors/parsing_errors.hpp"
 #include "errors/commandline_errors.hpp"
 #include "errors/internal_errors.hpp"
 #include "misc/console_colors.hpp"
+
 #include <iostream>
 
 void display_target_line(size_t char_pos, const std::string& line);
 
-void display_error_context(const std::string& filename, int target_line_number, int char_pos);
+void display_error_context(const std::string& filename, size_t target_line_number, size_t char_pos);
 
 void display_commandline_error(const CommandLineError& err);
